reorganize tests for mars-rover-health
