+++ contracts/health/src/lib.rs
@@ -2,4 +2,6 @@ pub mod compute;
 pub mod contract;
 pub mod querier;
 pub mod state;
+#[cfg(test)]
+mod tests;
 pub mod update_config;
