+++ contracts/health/src/tests/helpers/defaults.rs
