+++ contracts/health/src/tests/mod.rs
@@ -1,0 +1,7 @@
+mod health_state;
+mod health_values;
+mod helpers;
+mod hls;
+mod instantiate;
+mod liquidation_pricing;
+mod update_config;
